requirements.txt
streamlit>=1.39
python-dotenv>=1.0
google-genai>=1.38.0
google-generativeai>=0.8.5
Pillow>=10
tqdm>=4.66
